+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 ===========================
 
 
+## [2.2.1](https://github.com/buession/buession-security/releases/tag/v2.2.1) (2023-xx-xx)
+
+### 🔨依赖升级
+
+- [依赖库版本升级和安全漏洞修复](https://github.com/buession/buession-parent/releases/tag/v2.2.1)
+- **buession-geoip：** 升级 IP 库
+
+
+### 🔔 变化
+- **buession-web：** Servlet AbstractHandlerExceptionResolver doResolve 方法支持接收 handler
+
+
+---
+
+
 ## [2.2.0](https://github.com/buession/buession-security/releases/tag/v2.2.0) (2023-03-10)
 
 ### 🔨依赖升级